# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ORVELL PULSE Wholesale ERP Backend API",
        "description": "Production REST API documentation for ORVELL PULSE Wholesale ERP system with Spatie RBAC, multi-company isolation, inventory tracking, cashier settlement, and automated reporting.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default API Host"
        }
    ],
    "paths": {
...
        "/api/bank-deposits/create": {
            "post": {
                "tags": [
                    "Bank Deposits"
                ],
                "summary": "Record cashier bank deposit of counter cash",
                "description": "Record Bank Deposit (Cashier / Admin)",
                "operationId": "ec3e9918bbf1b1f4bebd6341e37ef874",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "bank_name"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 500
                                    },
                                    "bank_name": {
                                        "type": "string",
                                        "example": "GCB Bank"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1029384756"
                                    },
                                    "reference_number": {
                                        "type": "string",
                                        "example": "DEP-2026-001"
                                    },
                                    "deposit_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-09"
                                    },
                                    "deposit_slip_image": {
                                        "type": "string",
                                        "example": "https://storage.orvell.com/slips/slip1.jpg"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Morning till cash deposit"
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bank deposit recorded successfully"
                    },
                    "403": {
                        "description": "Unauthorized. Only authorized Cashiers, Managers, and Admins can record bank deposits"
                    },
                    "422": {
                        "description": "Validation error or duplicate reference"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bank-deposits/list": {
            "get": {
                "tags": [
                    "Bank Deposits"
                ],
                "summary": "List filtered bank deposits for authenticated company",
                "description": "List Bank Deposits",
                "operationId": "5806ca20938a6f892246294bc1477e39",
                "parameters": [
                    {
                        "name": "bank_name",
                        "in": "query",
                        "description": "Filter by bank name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "GCB Bank"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-08-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter end date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-08-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank deposits list retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bank-deposits/summary": {
            "get": {
                "tags": [
                    "Bank Deposits"
                ],
                "summary": "Get bank deposit summary aggregated by bank for date range",
                "description": "Bank Deposits Period Summary",
                "operationId": "bc3b1d9e6f15d274364d7066f728ae2b",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Summary start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-08-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Summary end date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-08-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank deposit summary retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum personal access token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}